requirement.txt
audio-recorder-streamlit==0.0.9
google-ai-generativelanguage==0.6.4
streamlit==1.29.0

